refactor(trackerless-network): User id as Ethereum address (#1830)
Handle the user id field similarly to other id fields: internally it is a string, so that we can use it e.g. as a map key. Before this PR it was stored as a byte array.

The protocol doesn't change: there it is transferred as a byte array, similarly to `NodeID` and others.

The user id is now annotated as `EthereumAddress` as it is that kind of string in practice. It would be possible to have it just as a branded string if we'd want to be more abstract about the content of the strings. (At `trackerless-network` level we don't care what kind of bytes it contains, it is just an ID of a user).