feat(trackerless-network): Remove `NameDirectory` (#1842)
Remove `NameDirectory` functionality. It provided human-readable names for nodes and contract addresses.

It is no longer relevant functionality as most of the node ids are unknown (as the network is decentralized). We could keep it for the contract address decoding. But as contract addresses change, it is maybe easier to debug potential problems if we output the actual contract addresses instead of using some static mapping which may be out-of-date.