feat(broker): [NET-981] Operator plugin accepts proxy connections (#1734)
Set `acceptProxyConnections` client config option to `true` by default when running `operator` plugin. If the client configuration contains an explicit `false` value, the Broker initialization throws an error.

## Implementation

There is a new feature in Broker which allows plugins to define config defaults for the client. A plugin can implement `getClientConfig()` method to define a list of config options which are requited for that plugin. The method checks whether there is a user-defined config value for that option and sets a value if there is no user-defined value. Broker throws an error If there is a different user-defined value (e.g. `false` when a plugin requires `true`).

## Refactoring

As the config for `StreamrClient` needs info from plugins, it is no longer possible to pass `StreamrClient` instance to `Plugin` constructor. Now the `StreamrClient` is a parameter of `Plugin#start` method instead.

## Future improvements

The list returned by `getClientConfig()` is not type-safe. The config option is a string path and value is `any`. It could be possible to add some type-safe there.