fix(network): Fix integration/RandomGraphNode-Layer1Node.test.ts issue with random IDs (#1848)
## Summary

Fixed issue with random IDs in
`integration/RandomGraphNode-Layer1Node.test.ts`.
The test had an assumption that in a stream graph of 5 nodes 4
handshakes would always be required per node to construct a complete
graph. This holds true with a certain sets of nodeIds. However, with
randomized IDs some cases will require 5 handshakes to construct a
complete graph.

The `SuccessListener`-logic was removed from the test and the affected
test cases now use `waitForCondition` to wait for the graph to be
complete.