test(trackerless-network): Unit-testable `NetworkNode` (#1851)
Make `NetworkNode` unit-testable.

## Changes

Change the visibility `NetworkNode` constructor from _public_ to _internal_. Add new `createNetworkNode` method to the public API for creating `NetworkNode` instances. Also change the constructor of `NetworkNode` so that it takes in `NetworkStack` dependency instead of creating it internally in the constructor.

These changes allow us to mock the dependencies of `NetworkNode` and thus create e.g. a unit test for it. The first unit test is implemented in https://github.com/streamr-dev/network/pull/1839.