feat(broker): Get `theGraphUrl` from client (#1859)
Add new internal `getConfig()` to `StreamrClient`. Use that method in `OperatorPlugin` to get the url of The Graph. 

The config object always contains a value: if user has not specified any value (in `StreamrClient` construtor / in Broker's config file) the value is the production default value.

Note that the current production default (`https://api.thegraph.com/subgraphs/name/streamr-dev/streams` in `config.schema.json` at line 278) value doesn't work as that is not the mono subgraph URL. We have not yet deployed the mono subgraph to production, but when we do and update the correct value to `config.schema.json,` Broker starts to use it automatically.