NET-1000: InspectRandomNodeService (#1773)
## Summary

Implement `InspectRandomNodeService` which on an interval selects a
random operator's node for inspection. The random operator is picked
from the set of sponsorships being worked on by "my" operator's node.

Full description of service is in the working document.

## Limitations and future improvements

- Creating the service (and starting it) is commented out in
`OperatorPlugin`. It will be a lot of code re-organization to get it
properly initialized so I think that is better served by a 2nd PR after
this one.
- There are some rare edge cases that could use unit tests (notice the
leftover TODO for this)
- Can we do the `operatorCount >= 2` filtering already in GraphQL side
(minor optimization)?
- We will need to extract part of the inspection logic for
`VoteOnSuspectNodeService` later (see NET-1059)
- Break up `InspectRandomNodeService` functions into their own files?