feat(broker) [ETH-552]: add leader check to `VoteOnSuspectNodeService` (#1737)
## Summary

Add leader check to service `VoteOnSuspectNodeService`. Only the leader
of an operator fleet state will inspect a flagged node and vote
accordingly.

## Changes

- Helper class `VoteOnSuspectNodeHelper` is passed to constructor
instead of created inside the constructor of the service. This makes the
service easier to test and avoids usage of hacky private readonly field
replacements.
- Add unit tests for the service, especially ones focusing on the effect
of being a leader node (and not being one).
- Clean up integration tests for the service.

## Limitations and future improvements

- Inspection is still not implemented. Placeholder value of `kick` is
used instead.
- Removed end-to-end test, important to cover this in upcoming
end-to-end test related to inspection. NET-1071