fix(network): Fix joining on streams when all entry points queried from DHT are offline. (#1840)
## Summary

Fix bug where if all stream entrypoints stored in the DHT are offline,
the stream would be inaccessible until the stored items' TTLs run out.

## Additional changes
- `joinDHT()` operation no longer throws on time outs. Instead a rejoin
is attempted as long as the new `doRejoin` is `true` (it is true by
default). Previously this kind of rejoining was sometimes an infinite
loop.
- layer1 joins only do DHT rejoins in cases where the stream entry
points are explicitly set.
- the scheduler of `PeerDiscovery#getClosestPeersFromBucket` invocation
is only started once.