fix(client): Invalidate permission cache after publish failure (#1885)
## Background

When a user tries to publish message, we always check whether the user has a publish permission for the stream. The result of that check is cached so that subsequent checks are super-fast. 

We should also invalidate the cache if the publishing fails, so that if someone grants the permission we don't cache the invalid `false` value (and therefore throw the error when there is actually a permission granted).

## Change

Now we invalidate cache in `MessageFactory.ts`. 

Fixed also a test case, which covered this use case, but it had incorrect await logic and therefore it didn't catch the bug.