feat(network): Advertised Websocket  server host (#1888)
## Summary

Added support for configuring an advertised websocket host name. This
will be needed in the docker environment and likely in cases where nodes
are ran behind a reverse-proxy or load balancer.

## Changes

- The `websocketHost` configuration is now used to advertise a host name
that the machine it self might not know of.

- WebSocket servers are always started in `0.0.0.0`

## Limitations and future improvements

The configuration of `externalIp` (for webrtc) and the advertised
`websocketHost` could be unified in some way, but it will require
thought on how to do it exactly. The details of starting websocket
servers will change after the auto-certifier is completed.

## Checklist before requesting a review

- [ ] Is this a breaking change? If it is, be clear in summary.
- [ ] Read through code myself one more time.
- [ ] Make sure any and all `TODO` comments left behind are meant to be
left in.
- [ ] Has reasonable passing test coverage?
- [ ] Updated changelog if applicable.
- [ ] Updated documentation if applicable.