feat(DHT): External store request (#1905)
## Summary

DhtNode now support ExternalStoreRequests. This makes it possible to
more reliably store data to the DHT while a nodes `joinDht()` operation
is still ongoing. By default `DhtNode#storeDataToDht` will elect to use
the external option if its join is still ongoing.

## Future improvements

- refactor external `findData` to use a similar pattern where the
DhtNode detects if it should use the external method instead of leaving
it to the higher layer.