fix(dht): [NET-1099] Incorrect event payload `SortedContactList#onContactRemoved` (#1916)
The event payload of `contactRemoved` was incorrect when it is emitted from `SortedContactList#addContact`. The first argument should be the removed contact but is the added contact.

That parameter is currently not used in any of the event listeners. Maybe good to keep the parameter for consistency, as similar parameter is also `newContact` event.