fix(dht): [NET-1100] Incorrect item is removed from contact list (#1917)
Incorrect item was removed when `SortedContactList#removeContact` or `RandomContactList#removeContact` was called.

The methods used referertial equality, which causes it to typically remove the last index from the array (as `indexOf` returns `-1` and we spliced with that index).

Improved the test coverage of the unit tests and simplified the state handling in the tests. The tests new create new ids `(PeerID.fromValue(descriptor.kademliaId)` so that we use different object references in the list vs. in the calls and thus find any bugs related to referential equality.

Also changed some usages of `splice` to `slice`. In those cases the intention is to return a sub-array of array, and not remove n items from the array. The `splice` worked in practice as it returns the items which it removes. But the current approach is more straightforward.