fix(trackerless-network): [NET-1104] Duplicate entry points in `joinStream` (#1920)
Fixed a bug in `StreamrNode#joinStream`: it put duplicate entry points to the array (same items twice). Also fixed  `layer1.joinDht` call to use all entry points, not just user-defined entry points (see open questions whether that is the correct fix)

Also fixed retry-parameter for `layer1.joinDht`: as a call to `joinStream` is always the first join to the stream, it is good to have it retry it until it succeeds (or timeouts).

Also fixed a flaky test.