refactor(trackerless-network): Divide `StreamrNode#joinStream` (#1927)
Divided the logic in `StreamrNode#joinStream` so that the method does the state management synchronously and the starting of layers is asynchrously (inside `setImmediate` call). Earlier the method return a `Promise`, but it was important that part of the method was called in the same tick.

As state management is no is now synchronous operation we can ensure that `this.streams` is always in consistent state (i.e. it contains all streams which we have joined or set proxies for, but other streams). After this change a call to `NetworkNode#joinStreamPart` always puts the internal state to the correct state (a stream joined). And in `StreamrNode#broadcast` we can just call `joinStream` without error-prone asynchronous management.

There is now also an additional check which ensures that `leaveStream()` has not been called before the background task starts.

The downside of this refactoring is that we can't no longer know when `StreamrNode#joinStream` operation has completed. But external components don't actually need to care about that internal state. Also the `StreamrNode` is never accessed directly by instead it is used via the `NetworkStack` class which already hide this internal state.

Also a fix for `DhtNode`: if the node has not been started, a call to `stop()` is a no-op. And increased timeout in `joining-streams-on-offline-peers.test.ts` to avoid flakiness.

## Future improvements
- rename `joinStream`/`leaveStream` to `joinStreamPart`/`leaveStreamPart` (and other similar methods?)
- rename `StreamrNode` class?
- maybe refactor `StreamrNodeType` to just a simple `proxied` boolean
- remove obsolete `IStreamNode` methods (e.g. `getTargetNeighborIds` and listeners?), and maybe the whole interface as `ProxyStreamConnectionClient` is actually not a node (may need small refactoring `StreamrNode.streams` data structure)`