refactor(dht): `Remote` doesn't implement `IContact` (#1934)
Refactored `Remote` to not to implement `IContact` as it is a RPC interface for a peer, not a peer itself. Also removed the `IContact` interface as there it is no longer needed.

Almost all `ContactList`s operate on `DhtPeer` instances. Some operate on Contact instances (a wrapper to any object that has a peer description). The `RoutingSession` was the exceptional case: it operated on `Remote` objects. Now it uses s custom `Contact` object which encloses also a router. This way the separation between a contact and the RPC interface is clearer.

## Related changes

- Moved `Events` and `ContactState` into `ContactList.ts` as those classes/interfaces are always operated via `ContactList`
- removed obsolete `RoutingSession#getClosestContacts` method