fix(network): NET-1017 fix flaky websocket-full-network test (#1946)
## Summary

Increased the the expected number of stream connections to be at least 4
before starting propagation in the test case. With this change the test
passed 100 times locally.

It used to be more difficult for all nodes to reach 4 connections but
now as we allow some nodes to have 5 connections 4 should always be
achievable for all nodes.