refactor(client): `getSigner()` type contains `Provider` (#1954)
Changed type of `StreamrClient#getSigner` to ensure that the signer also contains `Provider`. In `ethers` the `Signer` object has a field for provider, but that field is optional. 

In practice the signer we return is a `Wallet` or `JsonRpcSigner` and in both of those types the `Provider` is always non-null.