refactor(dht): Do not export internal classes (#1956)
Do not export these classes from `dht` package:
- `Contact`
- `SortedContactList`
- `PeerId`, `PeerIDKey`, `peerIdFromPeerDescriptor`
- `UUID`

We can consider these classes to be internal to the `dht` package as none of the classes is used in `dht`'s public API.

The `keyFromPeerDescriptor` is no longer needed outside of `dht` package as we can use the `binaryToHex` helper utility to convert `peerDescriptor.kademliaId` to `NodeId`. (That kind of direct conversion was already in used e.g. in `HandshakerServer`.

## Other changes

- removed `ILayer0#getNodeId()`. The same information is also available via `ILayer0#getPeerDescriptor()`
- changed `StreamPartHandshakeRequest` to use `uuid.v4()` directly instead of using `dht`'s UUID wrapper (all requests in the `dht` package also use `uuid.v4()` this way).
- removed  the `protoClasses` structure from exports
- fixed import statements which exported the classes/functions indirectly from `exports.ts`