refactor(trackerless-network): `streamPartId` in protocol messages (#1958)
Renamed `randomGraphId` field of these protocol messages to `streamPartId`:

- `StreamPartHandshakeRequest`, `InterleaveNotice`
- `LeaveStreamPartNotice`
- `NeighborUpdate`

That field must always contain a value of stream part id. 

Likewise renamed these config fields to `streamPartId`:
- `StrictRandomGraphNodeConfig#randomGraphId`
- `TemporaryConnectionRpcServerConfig#randomGraphId`
- `HandshakerConfig#randomGraphId`, `HandshakerServerConfig#randomGraphId`
- `NeighborUpdateManagerConfig#randomGraphId`, `NeighborUpdateManagerServerConfig#randomGraphId`
- `StreamNodeServerConfig#randomGraphId`
- `InspectorConfig#graphId`

The type of the field is `StreamPartID` instead of string. After these changes it is easy to see that each of these class operate on one stream part id. The custom type ensures that the value is always validated before it is used.

Also renamed `NeighborUpdateManagerConfig` -> `NeighborUpdateManagerServerConfig` in `NeighborUpdateManagerServer` (there is already a different interface with the same name in the `NeighborUpdateManager` class)