feat(dht, trackerless-network): Unify node id logging (#1967)
Unify how we log node's identity. Now the node's identity is always logged as the hexadecimal nodeId string, by using `keyFromPeerDescriptor`/`getNodeIdFromPeerDescriptor`.

In `ConnectionManager` and `ManagerConnection` classes we need to use a `keyOrUnknownFromPeerDescriptor` wrapper as the node's identity may be unavailable in some situations.

Before this PR some components logged the node name instead of node id. The node name is a local name for a node and should not be used when we log information about remote nodes. (See also NET-1130)

Also remove the logging of application's own node id from multiple `trace` level messages. The information is logged already when the node starts and is therefore redundant in those cases.

## Future improvements

- There are still some logging about the own identity in `Simulator` classes as we typically run multiple nodes when we use the simulator. In the future we could start to use `STREAMR_APPLICATION_ID` environment variable to scope the node's identity and remove the logging of node's own identity also from the `Simulator` classes.
- The same environment variable could be useful also in some end-to-end tests?