refactor(dht): `isBrowserEnvironment` (#1987)
Inverted `isNodeJS()` helper method to `isBrowserEnvironent()`. 

It is better to check the exceptional case (that we are in the browser environment), and assume that anything else is NodeJS (which is kind of approximation, e.g. Deno is not Node).

## Future improvements

The override is done only in client's webpack build, and not in `dht's` and `trackerless-network`'s karma.config.js (which override the `NodeWebRtcConnection` -> `BrowserWebRtcConnection`). Therefore browser tests currently use `NODEJS` as their node type instead of `BROWSER` node type. (NET-1143)