fix(DHT): NET-1140 fix critical memory leak in ListeningRpcCommunicator (#2030)
The `ListeningRpcCommunicator` had a critical bug where it set en event
listener to the `ITransport` passed to it, but it never cleaned it up.
This caused all `ListeningRpcCommunicators` created for
RecursiveFindSessions and RecursiveFindResponses to leak. Issue likely
also existed for `RandomGraphNodes` in the trackerless-network. If a
node a stream partition the `ListeningRpcCommunicator` created for it
would leak.

`ListeningRpcCommunicator` has now a new method called `destroy` which
ensures that the listener is cleaned up.
Also after a `RoutingSession` is completed the router will call stop on
in