refactor(dht): Remove `RouteMessageWrapper#previousPeer` (#2028)
Information about `RouteMessageWrapper#previousPeer` should be available from `routingPath` (it is the last item of the array).

Small change for `routingPath` handling in `RoutingSession`. The `routingPath` is now updated just before we send the message to the remote, so that we can easily analyze the value of `previousId` in the `RoutingSession` constructor.