feat(DHT): NET-1150 estimate rpc request timeouts (#2043)
## Summary

The rpc request timeouts are now estimated based on the expected
connection type that is used for the request. The estimation is only
done if a default timeout is not passed for the Remote super class.

- websocket client 5 seconds
- websocket server 7.5 seconds
- webrtc 20 seconds

## Future improvements

- By being able to detect if simulator is used and if the transport
layer is a DhtNode, the default timeout passing can be removed.
Currently components that use a Layer0Node or simulator as transport
need to explicitly set a default timeout to avoid the 20000 ms webrtc
timeout.
- On the other hand timing out connections that take over 15 seconds
might be good for the long run. Nodes will only connect to peers that
are near to them either geographically or in the kademlia structure.
However operations such as insocetion might suffer as a result. WebRTC
connections that take over 15 seconds will always fail inspection.
(Nodes that are running as operators should always have a WebSocket
server though)
- Rename `Remote` to `RpcRemote`?  (and `rpcTimeout` to `timeout`)