fix(DHT): WS server rejects connections if PeerDescriptor is invalid (#1904)
## Summary

Fixed bug where a WS server could accept multiple connections from the
same source if the `kademliaId` of the `PeerDescriptor` was invalid but
the server connection info was valid.

This fixes a problem where peers could assume an old peer that had left
the network is still online if the server of a new peer is same as the
old one.

This issue is not apparent in webrtc as we route signalling messages to
the correct peers based on kademliaId. New targetPeerDescriptor field
could be added for webrtc connection type as well but it is not needed

## Other changes

- Use `LatencyType.REAL` in simulations instead of `LatencyType.RANDOM`
to avoid problems with message ordering.

## Future improvements

- ManagedConnection refactor. Currently there are a lot of confusing
names and logic flow in the ManagedConnection class. Events such as
`bufferSentByOtherConnection` is quite confusing and could describe why
this is happening better. Handling of incoming and outgoing connections
is also quite confusing.
- Is there a better way to implement the newListener events in
ManagedConnection?
- The internal events are confusing in the ManagedConnection class,
would make sense to simplify the send() logic especially.
- Should IConnection Disconnected include the code and reason? Currently
the ManagedConnection does not do anything with the fields.

---------

Co-authored-by: Petri Savolainen <petri.savolainen@spaceify.net>