refactor(dht): `WebsocketServer` actions (#2067)
Refactor handling of `WebsocketServer` requests.

Also small change to the URL query parameter format:
- previously pattern like this `ws://foobar.com?connectivityRequest=true` 
- now the pattern like this: `ws://foobar.com?action=connectivityRequest`

## Open questions

There are three possible ways to open the connections: the default connection which starts with a handshake, and two actions `connectivityRequest`/`connectivityProbe`.  The idea is that when an action is executed, the server does a task and sends a response as a WebSocket message. After that message has been received the recipient is expected to close the socket connection.

But actually the `connectivityProbe` doesn't do anything. It just opens a connection which doesn't listen to messages, and doesn't send any message. 

Would it make sense to have just two modes/actions: `defaultRpc` (or something) and `connectivityRequest`. I.e. the `connectivityProbe` wouldn't be a mode as it doesn't do anything.