fix(dht, trackerless-network): Stop functionality in `DhtNode` and `StreamrNode` (#2069)
- In `DhtNode#stop` we stop `transport` only if the instance was constructed by this component (in `DhtNode#start`)
- In `NetworkStack#stop` we stop `layer0node` as we constructed the instance (moved it from `StreamrNode#stop`)
- In `StreamrNode#stop` we don't stop `layer0node` or `transport` as we didn't construct the instances

## Other changes

`NetworkStackStoppedDuringStart` faied in this PR. We remove the test, as we don't really support the feature,

If we want to support that feature in practice, the stop method should be able to wait until `start` completes. And vice versa. I.e. if `stop` is called before `start` completes, it fully starts the node, and postpones the actual teardown to happen just after the start completes.