fix(DHT): Fix external store storer PeerDescriptor (#2110)
## Summary

When storing data with the external store operation, the incorrect
PeerDescriptor was assigned for the storer. This could create scenarios
where the external node used to run the store operation would set itself
as the storer of data. When multiple request would be handled for the
same keys to store data the stored would be overwritten. Additionally,
the original storer would not be able to delete the data with its own
PeerDescriptor

## Future improvements

- Stop using setStreamPartEntryPoints in trackerless-network tests.
There could be an explicit test that tests the feature but by the
default behaviour should be used in all other tests