refactor: Unify throw type (#2114)
When we throw an error, it is now always an instance of `Error`.

We also use the `new` keyword when we construct new `Error` instances. That doesn't change the functionality, as it s ok to construct an Error without the `new` keyword. But unified style is easier to understand (as there is no specific reason to not to use the different construction style).