test(DHT): use random ids in store tests (#2112)
## Summary

Use random ids in store tests instead of static ids. This ensures that
in the long run more cases are tested.

## Future improvements

Could add a new test case that more explicitly validates that the data
is stored on the correct nodes. Now we always try to find the data after
storing which practically does the same but the checking could be
stricter.