feat(dht) Router send sad paths (#2087)
## Summary

When Router#send fails to pass a message to the network it will throw a
message. This means that RpcClients using a DhtNode as transport will be
able to detect failures faster instead of waiting for timeouts

## Future improvements

The RouteMessageAck proto generated message should include a optional
string field instead of a required string set to an empty string by
default.