refactor(dht): Rename `PeerDescriptor#kademliaId` (#2119)
## Background

When we'll start to use hybrid-DHT, the `PeerDescritor` id is no longer
a `kademliaId`. Instead the `PeerDescriptor` id is an abstract id, and
the `kademliaId` is the hash of that id.

## Changes

Renamed:
- `PeerDescriptor#kademliaId` -> `nodeId`
- `NetworkPeerDescriptor#id` -> `nodeId`
- `ClosestPeersRequest#kademliaId` -> `nodeId`

## Compatibility

Added minimalistic compatibility layer, which allows users to define and
use `NetworkPeerDescriptor` in the legacy format. See commit 767356e. We
can revert that commit after testnet/mainnet starts (as we instruct the
users to change their entry point configs anyway). It may be obsolete
even sooner if we implement these, and therefore won't have legacy
configs in practice:
-
https://linear.app/streamr/issue/NET-1133/update-config-wizard-to-create-10-configs
-
https://linear.app/streamr/issue/NET-1004/use-streamrconfig-for-streamr-client-production-values
    
## Future improvements

- rename `id` -> `nodeId` for each entry point in `@streamr/config`
- all `sourcePeer`/`targetPeer`/`destinationPeer` in `Router` (and
related) classes should be renamed to just `source`/`target`

---------

Co-authored-by: Eric Andrews <harbu88@hotmail.com>
Co-authored-by: Eric Andrews <eric.andrews@streamr.com>