fix(dht): Use `storeMaxTtl` config option (#2144)
Use the `storeMaxTtl` config option instead of hardcode `TTL_MAX`.

Before this PR the data storage TTL handling was done in two places: 
- in `LocalDataStore` using the hardcoded: it scheduled the auto-deletion to happen between `TTL_MIN` and `TTL_MAX`
- in `StoreRpcLocal` using the configured value:  it mutated the entry's `ttl` if it was greater than `storeMaxTtl`.

Now the TTL handling is done only in `LocalDataStore`. The`StoreRpcLocal` no longer mutates the entry. That way the `storeMaxTtl` option just controls how long this node keeps the entry, and the option doesn't affect the entries which are replicated to other nodes via this node.

Removed also hardcoded `TTL_MIN` limit. If we need the limit, we should add a `DhtNode` config option for that.