chore: temporary logging increase (testnet-one only) (#2141)
## Summary

Add some temporary `info` level logging to debug missing votes issue. Basically want to understand whether we are not getting `ReviewRequest` events from JSON RPC or whether we have bug in our code. 

This PR doesn't need to be ported to mainline dev branch 1.0 since it is a temporary thing.