fix(dht): Use `CUSTOM_GOING_AWAY` code in websocket close (#2147)
Added new `CUSTOM_GOING_AWAY` code.

In browser we can't send `GOING_AWAY`, because it is a reserved code. If we try to close a WebSocket with that code in a browser we get an error:
```
Uncaught (in promise) DOMException: Failed to execute ‘close’ on ‘WebSocket’: The code must be either 1000, or between 3000 and 4999. 1001 is neither.
```

We have two separate codes because browsers send `GOING_AWAY` code automatically when a tab is closed.