fix(DHT): fix setting staleness for data on new contacts (#2159)
## Summary

Fixed setting staleness for data stored in the DHT when new contacts
arrive.
If the local node is closest to data, set stale to false for the data
always. If the local node is not the closest and not within the closest
`redundancyFactor` nodes to the data set stale to true.