fix(dht): Update of data `stale` state in `StoreRpcLocal#onNewContact` (#2156)
Before this PR we incorrectly updated the data entry state to true when
`StoreRpcLocal#onNewContact` was called for the closest node.

This bug was introduced in
https://github.com/streamr-dev/network/pull/2137. In
`StoreRpcLocal#shouldReplicateDataToNewNode` there is a guard at line
`102`. After the refactoring we didn't have any special handling for the
guard case, and therefore we incorrectly called
`this.localDataStore.setStale(..., true)` if we returned from
`StoreRpcLocal#shouldReplicateDataToNewNode` because of the guard.

Now we have a method `replicateAndUpdateStaleStateIfClosest`. If the
local node is the closest node for the data:
- it always updates the stale state
- if the new node is quite close to the data (within
`redundancyFactor`), it replicates the data to that new node

## TODO

Does this change do the same thing, or something better than
https://github.com/streamr-dev/network/pull/2159? If something is
missing, we should include that into this PR or cherry-pick that PR 2159
to `streamr-1.0`.

## Open questions

Why we update our `stale` state only if we are the closest node for the
data?

## Other changes

Also renamed the `getNodesClosestToIdFromBucket` to
`getClosestNeighborsTo` in `StoreRpcLocal` as that is the name of
`PeerManager` method it calls.

---------

Co-authored-by: juslesan <santeri.juslenius@gmail.com>