fix(dht): Close connections after timeout (#2166)
## Summary

Fixed a critical issue where in cases where the `ManagedConnection's`
send time outs we do not close the connections in many cases. This was a
major issue in all cases that required signalling to form the connection
ie. requested ws connections and webrtc connections. In some cases the
websocket client connect would also be stuck in connecting state
forever. The clears up that issue as well.

## Future improvements

- Investigate why the `integration/store.test.ts` required an increased
afterEach timeout after the fix.
- Are both the closing and stopped fields needed in `ManagedConnection`