fix(trackerless-network): fix clean up (#2168)
## Summary

There were two problems with cleanup. After changes to destroy on StreamrNode the clean up on closing the process was not moved to NetworkStack. Additionally if a process was running multiple nodes, clean up was not called for all nodes.

## Future improvements

Could move exit listeners to their individual main components, for example if someone is running only a DhtNode it will not run the graveful leave operation on exit.