refactor(dht): [NET-1218] Use `NodeID` string internally (#2173)
In this PR we implement most of future improvements listed in https://github.com/streamr-dev/network/pull/2155. The node IDs are internally stored as strings instead of `PeerID` objects. This should optimize CPU usage as we do significantly less `PeerID.fromValue({})` calls and other unnecessary data conversion operations.

For data keys we still use `PeerIDKey`. We'll refactor that in a separate PR (maybe we should introduce a new `DataKey` branded string for it).

Also refactored `excludedNodeIDs` to be set `Set` instead of array. This should optimize the `SortedContactList#addContact` operation.


## Future improvements

- Use `sortedIndexBy` in `SortedContactList#removeContact`.
- See added TODO comments. E.g. we could remove areEqualNodeIds and just use === comparison. This is a good approach if compiler/`eslint` would warn if we'd incorrectly compare `NodeID` to `UInt8Array` with ===.
- Many classes could maybe pass `NodeID`s instead of `PeerDescriptor`s, if the receiver class is not interested about the contact information. Maybe e.g. `PeerDescriptor` could store `PeerDescriptor`s internally, but methods could get `NodeID`s as parameters / return types.