fix(network): Less optimistic interleaving (#2183)
## Summary

InterleaveNotice changed to a request. When interleaving the node will
hold its connection to close until the new one is opened. This will
reduce the likelihood of the numbers of stream connections falling
suddenly for nodes during churn.

Nodes will no longer accept handshakes from peers if an interleave
request ongoing to the requestor.

Nodes will no longer accept handshakes from peers if they have 4 or more
connections and too many ongoing interleave requests. This is done to
ensure that under heavy churn nodes always keep at least 2 connections.

This PR reduces the run time of the trackerless-network tests by 30-45
seconds depending on the run.

Observed reduced flakyness in client end-to-end tests

## Other changes

- Reduced the number of nodes in full-websocket-network.test.ts to 32.
The test was running with 150-200% cpu with 48 nodes.

##  Future improvements

- Investigate if nodes could hold atleast 3 connections at once.
Currently this causes performance issues as under churn most nodes will
reject incoming handshake requests causing the number of requests in the
stream in total to sky rocket.
- Perhaps nodes could temporarily accept more connections under heavy
churn? For example, N+2 overflowing connections could be disconnected in
the future (this could cause more handshakes to be fired over the
network)